7_seg_display/requirements.txt
pyserial==3.5
PyVirtualSerialPorts==1.0.1
